+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## 0.2.0 (DEV)
+## 0.2.0 (2022-09-10)
 
 - **BREAKING**: Remove generic type parameter 3 in `withPersistent()` call.
 - **BREAKING**: Remove generic type parameter 3 in `withPersistentMap()` call.
 - Fix: Improve type inference in `withPersistentMap()` call in some cases.
+- Fix: Update `peerDependencies` for `@cubux/storage-driver`.
 - Upd: Interface `WithPersistentOptions` is now generic with 3 optional type
   parameters.
 - Add: Options `serialize` for serialization before writing to driver and